Extracted CategoryListCell.
